AMP-136368 [langley] significant refactoring and simplification
